% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_functions.R
\name{sbc_add_col_sum}
\alias{sbc_add_col_sum}
\title{Agrega un renglon con totales a un dataframe}
\usage{
sbc_add_col_sum(x, sumLabel = "")
}
\description{
Las columnas no numericas contendran un caracter de longitud 0 o el texto del parametro charColText
}
