% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_functions.R
\name{sbc_camel}
\alias{sbc_camel}
\title{Convert the first letter of a text to upper case and the rest to lower case}
\usage{
sbc_camel(x)
}
\arguments{
\item{x}{text to be transofmed}
}
\description{
Transform every word in the string, also clean blank spaces at the begining and end
}
\examples{
camel("Helo world. my name is")

marcas15$marca <- map_chr(marcas15$marca ,camel) used to match case across dataframe brands
}
