% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_functions.R
\name{sbc_clen_outliers}
\alias{sbc_clen_outliers}
\title{Remove outliers from a vector
It make the value NA
Used to report numerical values without outliers
Just generate a new colum in the data set}
\usage{
sbc_clen_outliers(x)
}
\description{
Remove outliers from a vector
It make the value NA
Used to report numerical values without outliers
Just generate a new colum in the data set
}
