% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_functions.R
\name{sbc_compare_df_var_perc}
\alias{sbc_compare_df_var_perc}
\title{Compare one column on different data frames}
\usage{
sbc_compare_df_var_perc(
  x,
  var_name,
  transpos_ind = F,
  incidence_var = NULL,
  showCount = F
)
}
\arguments{
\item{x}{A list contaning the dataframes, names are used as name columns.}

\item{var_name}{variable to be compared}

\item{incidence_var}{if not null is used to calculate incedence (divide by length(unique(incidence_var)))}
}
\description{
Used to compare several years of a variable
}
