% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_functions.R
\name{sbc_get_col_sum}
\alias{sbc_get_col_sum}
\title{La función recibe una columna de un dataframe.
Solo regresa un numero o text, no aplica para dataframes
1) si la columna es numerica regresa la suma
2) si la columna es alfanumerica regresa un caracter de longitud 0 o lo que vega en el parametro charColText
Normalmente se llama desde la función 'sbc_sum_columns' para agregar un renglon con totales por columna}
\usage{
sbc_get_col_sum(x, sumLabel = "")
}
\description{
La función recibe una columna de un dataframe.
Solo regresa un numero o text, no aplica para dataframes
1) si la columna es numerica regresa la suma
2) si la columna es alfanumerica regresa un caracter de longitud 0 o lo que vega en el parametro charColText
Normalmente se llama desde la función 'sbc_sum_columns' para agregar un renglon con totales por columna
}
