% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_functions.R
\name{sbc_get_col_sum_row}
\alias{sbc_get_col_sum_row}
\title{get a row with sum of columns}
\usage{
sbc_get_col_sum_row(x, sumLabel = "")
}
\arguments{
\item{x}{dataset containig the data to report (\code{"hsv"} or \code{"cluster"})}

\item{sumLabel}{label of the row sum, if ommited it is an 0 length string}
}
\value{
Dataframe with a row of column sumns
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
plot_crayons()

}
