% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_functions.R
\name{sbc_get_num_col_fun}
\alias{sbc_get_num_col_fun}
\title{Aplica una función a una columna numerica
El resultado es agrupado por 'colVar'
Al final se agrega la columna ´A total´ que contien el resultado
de la función sin agrupar.
Se utiliza en sbc_summary_list para calcular la media y mediana}
\usage{
sbc_get_num_col_fun(x, colVar, col_name, FUN)
}
\description{
Aplica una función a una columna numerica
El resultado es agrupado por 'colVar'
Al final se agrega la columna ´A total´ que contien el resultado
de la función sin agrupar.
Se utiliza en sbc_summary_list para calcular la media y mediana
}
