% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_functions.R
\name{sbc_min_perc}
\alias{sbc_min_perc}
\title{Return rows with minimal value}
\usage{
sbc_min_perc(
  x,
  colName,
  minValue,
  sort = TRUE,
  desc = TRUE,
  otherLabel = "Otros"
)
}
\arguments{
\item{x}{The dataset}

\item{colName}{The column to compare}

\item{minValue}{The value to compare with}

\item{sort}{True if we need to sort the result on using de colName values}

\item{desc}{Order on descending order, if FALSE order in ascending order}
}
\description{
All rows that are less than the provided value are grouped in the last row
}
