% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_functions.R
\name{sbc_multi_column_to_df}
\alias{sbc_multi_column_to_df}
\title{Extrae un data.frame de varias columnas de una tabla}
\usage{
sbc_multi_column_to_df(x, idvar, pattern, outColumn)
}
\arguments{
\item{x}{source dataframe}

\item{idvar}{The key or primary key of the dataset}

\item{pattern}{Common text of the columns names to collect}

\item{outColumn}{The output column name}
}
\description{
Extrae un data.frame de varias columnas de una tabla
}
\examples{
sbc_multi_column_to_df(atsrv19_encuestas, "folio", "preferencia_medios", "medio")

Si tenemos las columnas preferencias_medios_01 a preferencias_medios_05
Haríamos lo siguiente: sbc_multi_column_to_df(atsrv19_encuestas, "folio", "preferencia_medios", "medio")
Como resultado tendríamos un df con las columnas folio y medio
}
