% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_report_list_v2.R
\name{sbc_report_list_v2}
\alias{sbc_report_list_v2}
\title{Regresa una lista con cruces de informaciín}
\usage{
sbc_report_list_v2(
  x,
  meta_data,
  groups,
  MARGIN = 2,
  show_global_count = T,
  show_section_count = T,
  show_section_percentage = F,
  total_first = T
)
}
\arguments{
\item{x}{El dataframe que se desea procesar}

\item{meta_data}{Los metadatos que describen a las variables que deseamos reportar
Generalmente se encuentran en un archivo de excel}

\item{groups}{Los grupos de columnas por los que queremos realizar los cruces}

\item{MARGIN}{2 - Suma 100 vertical , 1. Suma 100 horizonta}
}
\value{
Lista conteniendo los cruces
}
\description{
Regresa una lista con cruces de informaciín
}
