% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_functions.R
\name{sbc_summary_list}
\alias{sbc_summary_list}
\title{Lista resumen
Regresa una lista con varios dataframes, uno para cada valor del parámetro 'sectionVars'
Los dataframes contienen una columna para cada valor del parametro 'colVar'
Los renglones de los dataframes tienen un renglon para cada valor de 'sectionVars'
El primer dataset contiene la cuenta de casos}
\usage{
sbc_summary_list(
  x,
  colVar = NA,
  sectionVars,
  sectionNames = sectionVars,
  MARGIN = 2,
  showCount = FALSE,
  countLabel = "Casos",
  percentageLabel = "Porcentaje",
  totalLabel = "Total",
  showGlobalCount = T,
  showGlobalPercentage = T,
  showSectionCount = T,
  showSectionPercentage = T,
  globalCountSectionLabel = "Casos totales",
  na.rm = TRUE,
  incidenceVariable = NULL
)
}
\description{
Lista resumen
Regresa una lista con varios dataframes, uno para cada valor del parámetro 'sectionVars'
Los dataframes contienen una columna para cada valor del parametro 'colVar'
Los renglones de los dataframes tienen un renglon para cada valor de 'sectionVars'
El primer dataset contiene la cuenta de casos
}
