% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_functions.R
\docType{package}
\name{SbcRepTables}
\alias{SbcRepTables}
\title{SbcRepTables: A way to have "Summary Tables" for creating quick reports}
\description{
with Mardown or to export to excel- Allso include a couple of quick charts.
}
\section{Table functions}{

sbc_table_2_var_count
sbc_table_1_var_count
}

